\name{themejj-package}
\alias{themejj-package}
\alias{themejj}
\docType{package}
\title{
\packageTitle{themejj}
}
\description{
\packageDescription{themejj}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{themejj}
\packageIndices{themejj}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{themejj}

Maintainer: \packageMaintainer{themejj}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
