% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themejj.R
\name{themejj}
\alias{themejj}
\title{A ggplot theme I use for publications}
\usage{
themejj(base_size = 12, base_family = "Roboto Light", facet = FALSE)
}
\arguments{
\item{base_size}{The base size of the font; defaults to 12 pt}

\item{facet}{Show a rectangle around your plotting panel which can be useful if you have multiple facets,t hat is you us facet_grid(), facet_wrap(); defaults to FALSE (i.e., shows only x and y-axis line)}

\item{base_famile}{The font family; defaults to "Roboto Light". Note: you need to have the Roboto font isntalled and the fontextra package loaded [library(fontextra)] and the font installed [font_import(pattern="Roboto")] and [loadfonts()]}
}
\description{
This function applies a simple publication-ready theme
}
\examples{
themejj()
}
\keyword{ggplot2,}
\keyword{graphics}
\keyword{themes,}

