% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themejj.R
\name{themejj}
\alias{themejj}
\title{A ggplot theme: simple publication-ready theme}
\usage{
themejj(
  base_size = 10,
  base_family = .default_font,
  base_line_size = base_size/30,
  base_rect_size = base_size/30,
  facet = FALSE
)
}
\arguments{
\item{base_size}{The base size of the font; defaults to 10 pt}

\item{base_family}{base font family; defaults to "Roboto Condensed". Note: you need to have the Roboto font isntalled and the fontextra package loaded \code{library(fontextra)} and the font installed \code{library{font_import(pattern="Roboto")}} and \code{loadfonts()}. The package wil try to fetch the font.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{facet}{(optional, default is \code{FALSE}). If \code{TRUE} a rectangle around plotting panel is shown, useful if you use facetting (\code{facet_grid(), facet_wrap()}.}
}
\description{
A ggplot theme: simple publication-ready theme
}
\examples{
\donttest{
see here for more https://github.com/JanaJarecki/themejj/wiki
data(mtcars)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(size = 2, 
            fill="#83AEB2",
            color="#AEAEB7",
            alpha = .6) +
  geom_smooth(method = lm, alpha = .2, color = "black") +
  labs(title = 'Line Plot', caption = 'Source: mtcars',
      subtitle="Miles Per Gallon and Weight")
# ggsave("../img/lineplot.jpg", w = 100, h = 100, unit = 'mm')
  
data(iris)
ggplot(iris, aes(Sepal.Length, color = Species)) +
  geom_density() +
  scale_color_viridis_d() +
  labs(caption = 'Source: iris',
      subtitle="A subtitle")
data("midwest", package = "ggplot2")
ggplot(midwest, aes(x=area, y=poptotal)) + 
    geom_point(aes(col=state, size=popdensity), alpha = .3) + 
    geom_smooth(method="loess", se=F, color="black") + 
    xlim(c(0, 0.1)) + 
    ylim(c(0, 500000)) + 
    labs(subtitle="Area versus Population", 
         y="Population", 
         x="Area", 
         title="Scatterplot", 
         caption = "Source: midwest") +
    facet_wrap(~state) +
    scale_color_viridis_d() +
    themejj(facet = TRUE)
# see here for more \link[https://github.com/JanaJarecki/themejj/wiki]
}
}
\author{
Jana B. Jarecki
}
\keyword{ggplot2,}
\keyword{graphics}
\keyword{themes,}
